move modal to separate file and fic typos
